- Changed Completly the way that block position is detected by the board. - Fixed blocks falling throgh other blocks. - Implemented 2 methods that handle when blocks can move or not Right and Left. - Added line removal when the line 0 is complete.

+++ Tetris/Tetris/src/Tetris/Piece.java
@@ -371,7 +371,7 @@ public class Piece extends Node {
         this.boxAbsolutePoint = boxAbsolutePoint;
     }
 
-    public float getNumBox() {
+    public int getNumBox() {
         return numBox;
     }
